# apps/api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lintel Design OS API",
    "version": "0.1.0",
    "description": "Generated from the API's Zod schemas and route table (pnpm api:openapi). Problem `type` URNs: urn:lintel-design-os:problem:<code>"
  },
...
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/v1/design-objects/{objectId}": {
      "delete": {
        "operationId": "DesignVersions_removeObject",
        "summary": "Remove a design object from a DRAFT",
        "tags": [
          "DesignVersions"
        ],
        "parameters": [
          {
            "name": "objectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          },
          {
            "name": "X-Org",
            "in": "header",
            "required": false,
            "description": "The organization to act in (required when the caller belongs to several)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "description": "The resource's current strong ETag",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectDeleted"
                }
              }
            }
          },
          "default": {
            "description": "RFC 9457 problem",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-lintel-access": {
          "actions": [
            "design_version.author"
          ]
        },
        "x-lintel-identity": "INTERNAL"
      },
      "get": {
        "operationId": "DesignVersions_object",
        "summary": "Get a design object",
        "tags": [
          "DesignVersions"
        ],
        "parameters": [
          {
            "name": "objectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          },
          {
            "name": "X-Org",
            "in": "header",
            "required": false,
            "description": "The organization to act in (required when the caller belongs to several)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectResponse"
                }
              }
            }
          },
          "default": {
            "description": "RFC 9457 problem",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-lintel-access": {
          "kind": "authenticated"
        },
        "x-lintel-identity": "INTERNAL"
      },
      "patch": {
        "operationId": "DesignVersions_updateObject",
        "summary": "Update a design object of a DRAFT",
        "tags": [
          "DesignVersions"
        ],
        "parameters": [
          {
            "name": "objectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          },
          {
            "name": "X-Org",
            "in": "header",
            "required": false,
            "description": "The organization to act in (required when the caller belongs to several)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "description": "The resource's current strong ETag",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObjectUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectMutation"
                }
              }
            }
          },
          "default": {
            "description": "RFC 9457 problem",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-lintel-access": {
          "actions": [
            "design_version.author"
          ]
        },
        "x-lintel-identity": "INTERNAL"
      }
    },
    "/api/v1/design-versions/{versionId}": {
      "get": {
        "operationId": "DesignVersions_get",
        "summary": "Get a design version",
        "tags": [
          "DesignVersions"
        ],
        "parameters": [
          {
            "name": "versionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          },
          {
            "name": "X-Org",
            "in": "header",
            "required": false,
            "description": "The organization to act in (required when the caller belongs to several)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionResponse"
                }
              }
            }
          },
          "default": {
            "description": "RFC 9457 problem",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-lintel-access": {
          "kind": "authenticated"
        },
        "x-lintel-identity": "INTERNAL"
      },
      "patch": {
        "operationId": "DesignVersions_update",
        "summary": "Update a DRAFT version's pins, survey revision or metadata",
        "tags": [
          "DesignVersions"
        ],
        "parameters": [
          {
            "name": "versionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          },
          {
            "name": "X-Org",
            "in": "header",
            "required": false,
            "description": "The organization to act in (required when the caller belongs to several)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "description": "The resource's current strong ETag",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VersionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionResponse"
                }
              }
            }
          },
          "default": {
            "description": "RFC 9457 problem",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-lintel-access": {
          "actions": [
            "design_version.author"
          ]
        },
        "x-lintel-identity": "INTERNAL"
      }
    },
...
    "/api/v1/design-versions/{versionId}/objects": {
      "get": {
        "operationId": "DesignVersions_objects",
        "summary": "List a version's design objects",
        "tags": [
          "DesignVersions"
        ],
        "parameters": [
          {
            "name": "versionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 1024
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "default": 50,
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            }
          },
          {
            "name": "X-Org",
            "in": "header",
            "required": false,
            "description": "The organization to act in (required when the caller belongs to several)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DesignVersions_objects_200"
                }
              }
            }
          },
          "default": {
            "description": "RFC 9457 problem",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-lintel-access": {
          "kind": "authenticated"
        },
        "x-lintel-identity": "INTERNAL"
      },
      "post": {
        "operationId": "DesignVersions_addObject",
        "summary": "Add a design object to a DRAFT",
        "tags": [
          "DesignVersions"
        ],
        "parameters": [
          {
            "name": "versionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          },
          {
            "name": "X-Org",
            "in": "header",
            "required": false,
            "description": "The organization to act in (required when the caller belongs to several)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "description": "The resource's current strong ETag",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObjectInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectMutation"
                }
              }
            }
          },
          "default": {
            "description": "RFC 9457 problem",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-lintel-access": {
          "actions": [
            "design_version.author"
          ]
        },
        "x-lintel-identity": "INTERNAL"
      }
    },
...
    "/api/v1/design-versions/{versionId}/overrides": {
      "get": {
        "operationId": "DesignVersions_overrides",
        "summary": "List a version's relationship overrides",
        "tags": [
          "DesignVersions"
        ],
        "parameters": [
          {
            "name": "versionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          },
          {
            "name": "X-Org",
            "in": "header",
            "required": false,
            "description": "The organization to act in (required when the caller belongs to several)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverrideList"
                }
              }
            }
          },
          "default": {
            "description": "RFC 9457 problem",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-lintel-access": {
          "kind": "authenticated"
        },
        "x-lintel-identity": "INTERNAL"
      },
      "post": {
        "operationId": "DesignVersions_addOverride",
        "summary": "Add a relationship override to a DRAFT",
        "tags": [
          "DesignVersions"
        ],
        "parameters": [
          {
            "name": "versionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          },
          {
            "name": "X-Org",
            "in": "header",
            "required": false,
            "description": "The organization to act in (required when the caller belongs to several)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[!-~]{16,128}$"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "description": "The resource's current strong ETag",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OverrideCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverrideMutation"
                }
              }
            }
          },
          "default": {
            "description": "RFC 9457 problem",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-lintel-access": {
          "actions": [
            "design_version.author"
          ]
        },
        "x-lintel-identity": "INTERNAL"
      }
    },
    "/api/v1/design-versions/{versionId}/overrides/{overrideCode}": {
      "delete": {
        "operationId": "DesignVersions_removeOverride",
        "summary": "Remove a relationship override from a DRAFT",
        "tags": [
          "DesignVersions"
        ],
        "parameters": [
          {
            "name": "overrideCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64
            }
          },
          {
            "name": "versionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          },
          {
            "name": "X-Org",
            "in": "header",
            "required": false,
            "description": "The organization to act in (required when the caller belongs to several)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "description": "The resource's current strong ETag",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverrideDeleted"
                }
              }
            }
          },
          "default": {
            "description": "RFC 9457 problem",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-lintel-access": {
          "actions": [
            "design_version.author"
          ]
        },
        "x-lintel-identity": "INTERNAL"
      }
    },
...
    "/api/v1/validation-runs/{runId}": {
      "get": {
        "operationId": "DesignVersions_run",
        "summary": "Get a validation run",
        "tags": [
          "DesignVersions"
        ],
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          },
          {
            "name": "X-Org",
            "in": "header",
            "required": false,
            "description": "The organization to act in (required when the caller belongs to several)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationRunResponse"
                }
              }
            }
          },
          "default": {
            "description": "RFC 9457 problem",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-lintel-access": {
          "kind": "authenticated"
        },
        "x-lintel-identity": "INTERNAL"
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase Auth access token"
      }
    },
    "schemas": {
...
      "DesignVersions_objects_200": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ObjectResponse"
            }
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "items",
          "nextCursor"
        ],
        "additionalProperties": false
      },
...
      "ObjectDeleted": {
        "type": "object",
        "properties": {
          "deleted": {
            "type": "object",
            "properties": {
              "id": {
                "$ref": "#/components/schemas/Uuid"
              },
              "lineageId": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "lineageId"
            ],
            "additionalProperties": false
          },
          "designVersion": {
            "$ref": "#/components/schemas/VersionState"
          }
        },
        "required": [
          "deleted",
          "designVersion"
        ],
        "additionalProperties": false
      },
...
      "ObjectMutation": {
        "type": "object",
        "properties": {
          "object": {
            "$ref": "#/components/schemas/ObjectResponse"
          },
          "designVersion": {
            "$ref": "#/components/schemas/VersionState"
          }
        },
        "required": [
          "object",
          "designVersion"
        ],
        "additionalProperties": false
      },
...
      "ReasonInput": {
        "type": "string",
        "minLength": 1,
        "maxLength": 2000
      },
...
      "VersionUpdate": {
        "type": "object",
        "properties": {
          "roomRevisionId": {
            "$ref": "#/components/schemas/UuidInput"
          },
          "pins": {
            "$ref": "#/components/schemas/PinsInput"
          },
          "versionLabel": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ]
          },
          "changeReason": {
            "$ref": "#/components/schemas/ReasonInput"
          },
          "source": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          }
        },
        "additionalProperties": false
      }
    }
  }
}